refactor: 컴포넌트 단위로 디렉토리를 생성, 컴포넌트 명을 변경하다
Movie는 MovieItem, MovieContainer는 MovieList
