+++ index.html
@@ -8,6 +8,6 @@
   </head>
   <body>
     <movie-header></movie-header>
-    <movies-container></movies-container>
+    <movie-list></movie-list>
   </body>
 </html>
