+++ src/components/SearchInput/SearchInput.css
@@ -12,6 +12,6 @@
   width: 14px;
   border: 0;
   text-indent: -1000rem;
-  background: url('../image/search_button.png') transparent no-repeat 0 1px;
+  background: url('../../image/search_button.png') transparent no-repeat 0 1px;
   background-size: contain;
 }
