+++ src/components/Skeleton/Skeleton.css
