+++ src/components/Skeleton/Skeleton.js
