+++ src/components/index.js
@@ -1,13 +1,13 @@
-import Header from './Header';
-import MoviesContainer from './MoviesContainer';
-import SearchInput from './SearchInput';
-import Button from './Button';
-import Movie from './Movie';
-import Skeleton from './Skeleton';
+import Header from './Header/Header';
+import MovieList from './MovieList/MovieList';
+import SearchInput from './SearchInput/SearchInput';
+import Button from './Button/Button';
+import Movie from './MovieListItem/MovieListItem';
+import Skeleton from './Skeleton/Skeleton';
 
 customElements.define('movie-header', Header);
 customElements.define('search-input', SearchInput);
-customElements.define('movies-container', MoviesContainer);
+customElements.define('movie-list', MovieList);
 customElements.define('common-button', Button);
 customElements.define('movie-item', Movie);
 customElements.define('skeleton-item', Skeleton);
